src/HybridDb.Tests/Migrations/Commands/RemoveTableTests.cs
using System;
using HybridDb.Config;
using HybridDb.Migrations.Commands;
using Shouldly;
using Xunit;
using Xunit.Extensions;

namespace HybridDb.Tests.Migrations.Commands
{
    public class RemoveTableTests : HybridDbTests
    {
        [Fact]
        public void RemovesTable()
        {
            Execute(new CreateTable(new Table("Entities", new Column("Col1", typeof(string)))));

            Execute(new RemoveTable("Entities"));

            store.Database.QuerySchema().ShouldNotContainKey("Entities");
        }

        [Fact]
        public void DoesNotRequireReprojection()
        {
            new RemoveTable("Entities").RequiresReprojectionOf.ShouldBe(null);
        }

        [Fact]
        public void IsUnsafe()
        {
            new RemoveTable("Entities").Unsafe.ShouldBe(true);
        }
    }
}